+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/powah/GeneratorPlugin.kt
@@ -1,0 +1,12 @@
+package site.siredvin.peripheralworks.integrations.powah
+
+import dan200.computercraft.api.lua.LuaFunction
+import owmii.powah.lib.block.AbstractEnergyProvider
+
+class GeneratorPlugin(private val provider: AbstractEnergyProvider<*>) : BaseEnergyStoragePlugin(provider) {
+
+    @LuaFunction(mainThread = true)
+    fun getEnergyGeneration(): Long {
+        return provider.generation
+    }
+}